Move JSPromise to own file
